+++ iModelBridgeCore/ScalableMesh/STM/SMPointIndex.h
@@ -2,7 +2,7 @@
 //:>
 //:>     $Source: STM/SMPointIndex.h $
 //:>
-//:>  $Copyright: (c) 2017 Bentley Systems, Incorporated. All rights reserved. $
+//:>  $Copyright: (c) 2018 Bentley Systems, Incorporated. All rights reserved. $
 //:>
 //:>+--------------------------------------------------------------------------------------
 
@@ -581,7 +581,7 @@ public:
     virtual bool Destroy();
 
 
-    virtual void LoadIndexNodes(uint64_t& nLoaded, int level, bool headersOnly);
+    virtual void LoadIndexNodes(size_t& nLoaded, int level, bool headersOnly);
 
     uint32_t       GetNbObjects() const;
 
@@ -1580,7 +1580,7 @@ public:
     -----------------------------------------------------------------------------*/
     size_t              GetSplitTreshold() const;
 
-    void LoadIndexNodes(uint64_t& nLoaded, int level, bool headersOnly);
+    void LoadIndexNodes(size_t& nLoaded, int level, bool headersOnly);
     void SetGenerating(bool isGenerating)
         {
         m_isGenerating = isGenerating;
