+++ iModelBridgeCore/ScalableMesh/STM/SMPointIndex.hpp
@@ -2,7 +2,7 @@
 //:>
 //:>     $Source: STM/SMPointIndex.hpp $
 //:>
-//:>  $Copyright: (c) 2017 Bentley Systems, Incorporated. All rights reserved. $
+//:>  $Copyright: (c) 2018 Bentley Systems, Incorporated. All rights reserved. $
 //:>
 //:>+--------------------------------------------------------------------------------------
 
@@ -6617,7 +6617,7 @@ template<class POINT, class EXTENT> bool SMPointIndexNode<POINT, EXTENT>::Query 
     return digDown;
     }
 
-template<class POINT, class EXTENT> void SMPointIndexNode<POINT, EXTENT>::LoadIndexNodes(uint64_t& nLoaded, int level, bool headersOnly)
+template<class POINT, class EXTENT> void SMPointIndexNode<POINT, EXTENT>::LoadIndexNodes(size_t& nLoaded, int level, bool headersOnly)
 {
     HINVARIANTS;
 
@@ -6650,7 +6650,7 @@ template<class POINT, class EXTENT> void SMPointIndexNode<POINT, EXTENT>::LoadIn
         }
 }
 
-template<class POINT, class EXTENT> void SMPointIndex<POINT, EXTENT>::LoadIndexNodes(uint64_t& nLoaded, int level, bool headersOnly)
+template<class POINT, class EXTENT> void SMPointIndex<POINT, EXTENT>::LoadIndexNodes(size_t& nLoaded, int level, bool headersOnly)
 {
     if (m_pRootNode != NULL)
         {
